Mudei algumas coisas, e adicionei uma matriz
