+++ KAUA_MATHEUS/primeiro_bimestre/arquivoCompras.txt
@@ -1,2 +1,3 @@
-Qnt: 100 -- Preço: R$2.0 -- Desconto Total: R$10,00
-Qnt: 4 -- Preço: R$5.0 -- Desconto Total: R$0,00
+Qnt: 9 -- Preço: R$21.0 -- Desconto Total: R$0,00
+Qnt: 10 -- Preço: R$9.0 -- Desconto Total: R$0,00
+Qnt: 8 -- Preço: R$2.0 -- Desconto Total: R$0,00
